Performance improvements for bytes collections, decreased index size by setting the relative offset

+++ streaming/src/config.rs
@@ -26,7 +26,7 @@ pub struct PartitionConfig {
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct SegmentConfig {
-    pub size_bytes: u64,
-    pub messages_required_to_save: u64,
-    pub messages_buffer: u64,
+    pub size_bytes: u32,
+    pub messages_required_to_save: u32,
+    pub messages_buffer: u32,
 }
